Fixed bug in timeout type

+++ parity_core.py
@@ -9,6 +9,7 @@ import sys
 from rdkit import Chem
 from rdkit.Chem import AllChem
 from rdkit.Chem import Descriptors
+from rdkit.Chem.rdchem import EditableMol
 from rdkit import DataStructs
 from rdkit.Chem.Fingerprints import FingerprintMols
 from rdkit.Chem.rdmolfiles import MolFromMolFile
@@ -25,11 +26,12 @@ def generate_parity(sm_1,sm_2,reduce_timeout=1):
     try:
         mol_1 = MolFromMolFile(sm_1)
     except Exception as e:
-        mol_1=Chem.MolFromSmiles(sm_1)
+        mol_1 = Chem.MolFromSmiles(sm_1)
     try:
         mol_2 = MolFromMolFile(sm_2)
     except Exception:
-        mol_2=Chem.MolFromSmiles(sm_2)
+        mol_2 = Chem.MolFromSmiles(sm_2)
+
 
     ms = [mol_1,mol_2]
     sim_arr=[None,None,None,None]
@@ -44,11 +46,11 @@ def generate_parity(sm_1,sm_2,reduce_timeout=1):
     sim_score_best=None
     markush_best=None
 
-    mcs_graph_bondtypes_elements=MCS.FindMCS(ms,bondCompare=BondCompare.CompareOrder,atomCompare=AtomCompare.CompareElements,timeout=10/faster, ,completeRingsOnly=True)
+    mcs_graph_bondtypes_elements=MCS.FindMCS(ms,bondCompare=BondCompare.CompareOrder,atomCompare=AtomCompare.CompareElements,timeout=int(10/faster),completeRingsOnly=True)
     matches_best,sim_score_best,markush_best=generate_sim_score(mol_1,mol_2,mcs_graph_bondtypes_elements.smartsString)
 
 
-    mcs_graph_any_elements=MCS.FindMCS(ms,bondCompare=BondCompare.CompareAny,atomCompare=AtomCompare.CompareElements,timeout=10/faster,completeRingsOnly=True)
+    mcs_graph_any_elements=MCS.FindMCS(ms,bondCompare=BondCompare.CompareAny,atomCompare=AtomCompare.CompareElements,timeout=int(10/faster),completeRingsOnly=True)
     matches,sim_score,markush=generate_sim_score(mol_1,mol_2,mcs_graph_any_elements.smartsString)
 
     if sim_score>sim_score_best:
@@ -56,7 +58,7 @@ def generate_parity(sm_1,sm_2,reduce_timeout=1):
             sim_score_best=sim_score
             markush_best=markush
 
-    mcs_graph_any_any=MCS.FindMCS(ms,bondCompare=BondCompare.CompareAny,atomCompare=AtomCompare.CompareAny,timeout=30/faster,completeRingsOnly=True)
+    mcs_graph_any_any=MCS.FindMCS(ms,bondCompare=BondCompare.CompareAny,atomCompare=AtomCompare.CompareAny,timeout=int(30/faster),completeRingsOnly=True)
     matches,sim_score,markush=generate_sim_score(mol_1,mol_2,mcs_graph_any_any.smartsString)
     if sim_score>sim_score_best:
             matches_best=matches
